requirements.txt
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
mysql-connector-python==9.0.0
pycparser==2.22
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
requests==2.32.3
shiboken6==6.7.2
urllib3==2.2.2
